Move models to workspace sdk
